linux-stable.sh: Make kernelversion only ever output the version
There are times where exporting a variable or out of tree changes to the
kernel's main Makefile can cause extra output when just running
'make kernelversion'. To get around this, add the '-s' flag to make and
explicitly set CC and CROSS_COMPILE so that there should never be extra
output.

Closes: #6
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>